Undefined behaviour in six C23 <stdbit.h> functions, and the cast is next door
CBMC, on the first real run over the soundly-checkable set:

  FAILED lib/libc/stdbit/stdc_leading_ones.c:stdc_leading_ones_uc
      line 20 arithmetic overflow on signed shl in (signed int)x << offset

    unsigned int
    stdc_leading_ones_uc(unsigned char x)
    {
            const int offset = UINT_WIDTH - UCHAR_WIDTH;   /* 24 */

            return (__builtin_clz(~(x << offset)));
    }

x is unsigned char, so integer promotion makes it an int (C17 6.3.1.1p2 -
int represents every unsigned char value). For x >= 128, x << 24 is
255 * 2^24 = 0xFF000000 at the top, and INT_MAX is 0x7FFFFFFF. C17
6.5.7p4: if E1 has a signed type and E1 x 2^E2 is not representable in the
result type, THE BEHAVIOUR IS UNDEFINED.

The same for unsigned short at offset 16, for every x >= 32768.

Confirmed with a second instrument before touching anything, because one
tool saying UB is one tool's opinion:

  $ gcc -fsanitize=undefined -fno-sanitize-recover=all
  runtime error: left shift of 255 by 24 places cannot be represented
                 in type 'int'

Six functions across three files:

  stdc_leading_ones.c        _uc  _us
  stdc_leading_zeros.c       _uc  _us
  stdc_first_leading_one.c   _uc  _us

And the fix was already in the directory. stdc_first_leading_zero.c, same
author, same month, same shape:

  return (__builtin_clz(~(unsigned int)x << offset) + 1);
                         ^^^^^^^^^^^^^^

It casts. The other three do not. That file also guards clz(0) with
`if (x == UCHAR_MAX) return (0);`, and first_leading_one guards `x == 0`
the same way - so the zero case was thought about everywhere and the
promotion was not thought about in three files out of four. An oversight
with the correct form sitting next door, rather than a decision.

So: (unsigned int) on the shifted operand, and nothing else. Promoting to
unsigned makes the shift defined - C17 6.5.7p4 gives unsigned left shift a
value modulo 2^N - and the bit pattern is identical on any machine this
targets, which is why the bug never showed.

Checked, rather than asserted:

  - exhaustively, against an INDEPENDENT reference implementation that
    counts bits from the top in a loop, not against the old code: all 256
    unsigned char values and all 65,536 unsigned short values for each of
    the three operations, 197,376 cases, ZERO mismatches, under UBSan the
    whole way. The fix cannot have changed a result.
  - CBMC over lib/libc/stdbit afterwards: 70 of 70 functions PROVED, where
    three were FAILED.

This is upstream FreeBSD code (stdbit(3) is new in 15) and the bug is
upstream's, not this tree's.

+++ hbsd/src/lib/libc/stdbit/stdc_first_leading_one.c
@@ -15,7 +15,7 @@ stdc_first_leading_one_uc(unsigned char x)
 	if (x == 0)
 		return (0);
 
-	return (__builtin_clz(x << offset) + 1);
+	return (__builtin_clz((unsigned int)x << offset) + 1);
 }
 
 unsigned int
@@ -26,7 +26,7 @@ stdc_first_leading_one_us(unsigned short x)
 	if (x == 0)
 		return (0);
 
-	return (__builtin_clz(x << offset) + 1);
+	return (__builtin_clz((unsigned int)x << offset) + 1);
 }
 
 unsigned int
